Refactor bootstrapping script to log individual lines
This commit refactors the bootstrapping script to log individual lines
rather than blocks and other segmentations of code. This is a first step
in reducing the complexity of the script by allowing it to log
with the same logging functions rather than with complex nesting and
difficult to parse messages.

Bootstrapper changes:
- Make logging for bootstrap script consistent and go line by line (the
  UX is a work in progress, but this the intended direction)
- Polish linking steps
- Consolidate WSL2 checking logic
- Disable cargo package install for darwin
- Polish Rust install functions
- Remove neovim plugin step
- Split non-essential packages for darwin into a new file
- Remove Git user name step

Other changes:
- Retire tmux and neovim configs (and remove tmux and neovim from darwin
  base packages)
- Move neovim update plugins function from global update function
  to its own, independent function

Signed-off-by: Nick Gerace <nickagerace@gmail.com>